+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (Nov 12, 2022)
 
 - Add `--clean` option
+- Repository name validation
 - Remove workdir in docker file
 
 ## 0.1.0 (Nov 06, 2022)